+++ nengo/networks/tests/test_circularconv.py
@@ -3,7 +3,7 @@ import pytest
 
 import nengo
 from nengo.networks.circularconvolution import circconv, transform_in, transform_out
-from nengo.utils.numpy import rmse
+from nengo.utils.numpy import rms
 
 
 @pytest.mark.parametrize("invert_a", [True, False])
@@ -56,8 +56,8 @@ def test_input_magnitude(Simulator, seed, rng, dims=16, magnitude=10):
     with Simulator(model) as sim:
         sim.run(0.01)
 
-    error = rmse(result, sim.data[res_p][-1]) / (magnitude ** 2)
-    error_bad = rmse(result, sim.data[res_p_bad][-1]) / (magnitude ** 2)
+    error = rms(result - sim.data[res_p][-1]) / (magnitude ** 2)
+    error_bad = rms(result - sim.data[res_p_bad][-1]) / (magnitude ** 2)
 
     assert error < 0.1
     assert error_bad > 0.1
@@ -81,7 +81,7 @@ def test_neural_accuracy(Simulator, seed, rng, dims, neurons_per_product=128):
     with Simulator(model) as sim:
         sim.run(0.01)
 
-    error = rmse(result, sim.data[res_p][-1])
+    error = rms(result - sim.data[res_p][-1])
 
     assert error < 0.1